predict-sentiment-api/requirements.txt
fastapi==0.68.1
Werkzeug==2.0.1
uvicorn==0.15.0
oci==2.44.1
jsonpickle==2.0.0
pandas==1.3.2
requests
python-dotenv
nltk
